<ci>[onmi]:修改了cmdcallback相关的设置

+++ omniGKF_ctrl/include/omniGKF_ctrl/omni_gkf_usb.hpp
@@ -25,7 +25,8 @@ namespace omni_gkf
         ~OmniGKFUSB();
 
         void update();
-        void Set(uint8_t cmd, int16_t data);
+        // void Set(uint8_t cmd, int16_t data);
+        void Set(uint8_t cmd, float data);
         void haveData();
         bool isAvailable();
